Use separate cache keys for repository cache.
The repository cache should only depend on the contents of WORKSPACE and
Starlark files, but not on the OS and Bazel version,
cf. https://docs.bazel.build/versions/4.2.2/guide.html#the-repository-cache.
Right now the cache is so big that it’s always evicted.  Keying the repository
cache on WORKSPACE and Starlark file contents only should reduce the overall
cache size.