Use flag to selectively enable treat_warnings_as_errors.
We want to enable treat_warnings_as_errors only if we're building
the main repository (to isolate users from warnings), and only
for the targets in our repository.  Specifying a private flag
in .bazelrc seems to be the simplest way to achieve this.

This partially reverts commit 0ae8e08e15cd61ef6de98f0797b6e51e666ac371.